[refactor] exports as per review suggestion
