+++ src/components/index.jsx
@@ -2,3 +2,6 @@ export { Filters } from "./Filters";
 export { Footer } from "./Footer";
 export { Navbar } from "./Navbar";
 export { ScrollToTop } from "./ScrollToTop";
+export { CartCard } from "./CartCard";
+export { PriceSummary } from "./PriceSummary";
+export { ProductCard } from "./ProductCard";
